Extracted time_series_regression_head
